refactor: Move detection to a a package (revision 1)
