+++ pkg/analyzer/detection/errors.go
@@ -1,0 +1,13 @@
+package detection
+
+import "github.com/stackrox/scanner/pkg/commonerr"
+
+var (
+	// ErrUnsupported is the error that should be raised when an OS or package
+	// manager is not supported.
+	ErrUnsupported = commonerr.NewBadRequestError("worker: OS and/or package manager are not supported")
+
+	// ErrParentUnknown is the error that should be raised when a parent layer
+	// has yet to be processed for the current layer.
+	ErrParentUnknown = commonerr.NewBadRequestError("worker: parent layer is unknown, it must be processed first")
+)
